Move formatters to utils/formatters
