+++ src/pages/MyMove/Review/index.jsx
@@ -22,7 +22,7 @@ class Review extends Component {
     const { pages, pageKey, canMoveNext } = this.props;
 
     return (
-      <div className="review-move-container">
+      <div className={styles.reviewMoveContainer}>
         <WizardPage handleSubmit={noOp} pageList={pages} pageKey={pageKey} pageIsValid canMoveNext={canMoveNext}>
           <div className={`${styles.reviewMoveHeaderContainer} grid-row`}>
             <h2 className="tablet:grid-col-10" data-testid="review-move-header">
