+++ src/scenes/Review/Review.css
@@ -110,7 +110,3 @@
 .hhg-shipment-summary .not-implemented {
   display: none;
 }
-
-.review-move-container {
-  margin-top: -20px;
-}
