+++ src/shared/utils.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { get, includes, find, mapValues, capitalize } from 'lodash';
-import moment from 'moment';
 import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import faClock from '@fortawesome/fontawesome-free-solid/faClock';
 import faCheck from '@fortawesome/fontawesome-free-solid/faCheck';
@@ -155,22 +154,4 @@ export function dateSort(field, direction) {
       return Date.parse(a[`${field}`]) - Date.parse(b[`${field}`]);
     };
   }
-}
-
-// Format dates for customer app (ex. 25 Dec 2020)
-export function formatCustomerDate(date) {
-  return moment(date).format('DD MMM YYYY');
-}
-
-// Format orders type (ex: PERMANENT_CHANGE_OF_STATION => Permanent change of station)
-export function formatOrderType(orderType) {
-  return orderType
-    .split('_')
-    .map((str, i) => {
-      if (i === 0) {
-        return str[0] + str.slice(1).toLowerCase();
-      }
-      return str.toLowerCase();
-    })
-    .join(' ');
 }
