+++ src/utils/formatters.js
@@ -1,0 +1,19 @@
+import moment from 'moment';
+
+// Format orders type (ex: PERMANENT_CHANGE_OF_STATION => Permanent change of station)
+export function formatOrderType(orderType) {
+  return orderType
+    .split('_')
+    .map((str, i) => {
+      if (i === 0) {
+        return str[0] + str.slice(1).toLowerCase();
+      }
+      return str.toLowerCase();
+    })
+    .join(' ');
+}
+
+// Format dates for customer app (ex. 25 Dec 2020)
+export function formatCustomerDate(date) {
+  return moment(date).format('DD MMM YYYY');
+}
